--- a/sparse/tests/test_compressed_convert.py
+++ b/sparse/tests/test_compressed_convert.py
@@ -1,27 +1,62 @@
 import sparse
 import pytest
 import numpy as np
+from numba.typed import List
 
 import sparse._compressed.convert as convert
 from sparse._utils import assert_eq
 
 
-# @pytest.mark.parametrize(
-# )
-def test_convert_to_flat():
-    ...
+def make_increments(shape):
+    inds = [np.arange(1, a // 2) for a in shape]
+    shape_bins = convert.transform_shape(np.asarray(shape))
+    increments = List([inds[i] * shape_bins[i] for i in range(len(shape))])
+    return increments
 
 
-# @pytest.mark.parametrize(
-# )
-def test_compute_flat():
-    ...
-    # assert_eq(x[index], s[index])
+@pytest.mark.parametrize(
+    "shape, expected_p100",
+    [
+        [(5, 6, 7, 8, 9), np.array([3610])],
+        [
+            (13, 12, 12, 9, 7),
+            np.array([9899, 12244, 19923, 28043, 30388, 38067, 46187, 48532]),
+        ],
+        [
+            (12, 15, 7, 14, 9),
+            np.array(
+                [
+                    14248,
+                    16166,
+                    18786,
+                    29278,
+                    31898,
+                    41754,
+                    44380,
+                    54866,
+                    57486,
+                    68050,
+                    69968,
+                ]
+            ),
+        ],
+        [(9, 9, 12, 7, 12), np.array([10177, 12193, 20257, 28321, 30337])],
+    ],
+)
+def test_compute_flat(shape, expected_p100):
+    increments = make_increments(shape)
+    operations = np.prod(
+        [inc.shape[0] for inc in increments[:-1]], dtype=increments[0].dtype
+    )
+    cols = np.empty(increments[-1].size * operations, dtype=increments[0].dtype)
+
+    assert_eq(convert.compute_flat(increments, cols, operations)[::100], expected_p100)
 
 
 @pytest.mark.parametrize(
     "shape, expected_shape",
     [
+        [(5, 6, 7, 8, 9), np.array([3024, 504, 72, 9, 1])],
         [(13, 12, 12, 9, 7), np.array([9072, 756, 63, 7, 1])],
         [(12, 15, 7, 14, 9), np.array([13230, 882, 126, 9, 1])],
         [