--- a/sparse/tests/test_compressed_convert.py
+++ b/sparse/tests/test_compressed_convert.py
@@ -57,4 +57,6 @@ def test_compute_flat(shape, expected_subsample, subsample):
     ],
 )
 def test_transform_shape(shape, expected_shape):
-    assert_eq(convert.transform_shape(np.asarray(shape)), expected_shape)
+    assert_eq(
+        convert.transform_shape(np.asarray(shape)), expected_shape, compare_dtype=False
+    )