--- a/sparse/tests/test_compressed_convert.py
+++ b/sparse/tests/test_compressed_convert.py
@@ -31,7 +31,7 @@ def test_compute_flat(shape, expected_subsample, subsample):
     increments = make_increments(shape)
     dtype = increments[0].dtype
     operations = np.prod([inc.shape[0] for inc in increments[:-1]], dtype=dtype)
-    cols = np.empty(increments[-1].size * operations, dtype=dtype)
+    cols = np.tile(increments[-1], operations)
 
     assert_eq(
         convert.compute_flat(increments, cols, operations)[::subsample],