@hameerabbasi 2022-11-22
I guess there's an astype or an explicit dtype missing somewhere, from a quick glance at the CI logs.