@jamestwebber 2022-11-22
I realized I could test for `convert_to_flat`, which is probably redundant with the test for `compute_flat`. I don't think it would hurt to remove the latter.